fix(cli): emit snapshots for approval allowlist JSON no-ops
